+++ R/styles.R
@@ -325,7 +325,23 @@ color <- function(x, i = NULL, j = NULL, color, part = "body", source = j ){
 #' @param i rows selection
 #' @param j columns selection
 #' @param part partname of the table (one of 'all', 'body', 'header', 'footer')
-#' @param fontname string value, the font name.
+#' @param fontname single character value. With Word and PowerPoint output, the value specifies the font to
+#' be used to format characters in the Unicode range (U+0000-U+007F).
+#' @param cs.family Optional font to be used to format
+#' characters in a complex script Unicode range. For example, Arabic
+#' text might be displayed using the "Arial Unicode MS" font.
+#' Used only with Word and PowerPoint outputs. Its default value is the value
+#' of `fontname`.
+#' @param eastasia.family optional font to be used to
+#' format characters in an East Asian Unicode range. For example,
+#' Japanese text might be displayed using the "MS Mincho" font.
+#' Used only with Word and PowerPoint outputs. Its default value is the value
+#' of `fontname`.
+#' @param hansi.family optional. Specifies the font to be used to format
+#' characters in a Unicode range which does not fall into one of the
+#' other categories.
+#' Used only with Word and PowerPoint outputs. Its default value is the value
+#' of `fontname`.
 #' @family sugar functions for table style
 #' @examples
 #' require("gdtools")
@@ -342,7 +358,7 @@ color <- function(x, i = NULL, j = NULL, color, part = "body", source = j ){
 #' \if{html}{\figure{fig_font_1.png}{options: width=70\%}}
 #'
 #' \if{html}{\figure{fig_font_2.png}{options: width=70\%}}
-font <- function(x, i = NULL, j = NULL, fontname, part = "body" ){
+font <- function(x, i = NULL, j = NULL, fontname, part = "body", cs.family = fontname, hansi.family = fontname, eastasia.family = fontname ){
 
   if( !inherits(x, "flextable") ) stop("font supports only flextable objects.")
   part <- match.arg(part, c("all", "body", "header", "footer"), several.ok = FALSE )
@@ -362,6 +378,9 @@ font <- function(x, i = NULL, j = NULL, fontname, part = "body" ){
   j <- get_columns_id(x[[part]], j )
 
   x[[part]]$styles$text[i, j, "font.family"] <- fontname
+  x[[part]]$styles$text[i, j, "cs.family"] <- cs.family
+  x[[part]]$styles$text[i, j, "hansi.family"] <- hansi.family
+  x[[part]]$styles$text[i, j, "eastasia.family"] <- eastasia.family
   x
 }