+++ R/html_str.R
@@ -1,8 +1,7 @@
 caption_html_str <- function(x, bookdown = FALSE){
   tab_props <- opts_current_table()
-
   # caption "bookmark"
-  bookdown_ref_label <- ref_label()
+  bookdown_ref_label <- ref_label(tab_props$tab.lp)
   if(bookdown && !is.null(x$caption$autonum$bookmark)){
     bookdown_ref_label <- paste0("(\\#", x$caption$autonum$seq_id, ":",
                                  x$caption$autonum$bookmark, ")")
@@ -29,7 +28,7 @@ caption_html_str <- function(x, bookdown = FALSE){
   }
   caption
 }
-html_str <- function(x, ft.align = NULL, class = "tabwid", caption = "", shadow = TRUE){
+html_str <- function(x, ft.align = NULL, class = "tabwid", caption = "", shadow = TRUE, topcaption = TRUE){
 
   fixed_layout <- x$properties$layout %in% "fixed"
   if(!fixed_layout){
@@ -45,8 +44,10 @@ html_str <- function(x, ft.align = NULL, class = "tabwid", caption = "", shadow 
   classname <- gsub("(^[[:alnum:]]+)(.*)$", "cl-\\1", classname)
   tabcss <- paste0(".", classname, "{", tabcss, "}")
 
-  codes <- sprintf("<style>%s%s%s</style><table class='%s'>%s%s</table>",
-          tabcss, codes$css, flextable_global$defaults$extra_css, classname, caption, codes$html)
+  if(topcaption) str <- paste0(caption, codes$html)
+  else str <- paste0(codes$html, caption)
+  codes <- sprintf("<style>%s%s%s</style><table class='%s'>%s</table>",
+          tabcss, codes$css, flextable_global$defaults$extra_css, classname, str)
 
   if( is.null(ft.align) ) ft.align <- "center"